[new_feature]  重构代码 see http://rb.corp.taobao.com/r/66573 [1] 将处理消息逻辑移到handle_task_helper中 [2] 添加查询进度操作 [3] 添加单元测试（这部分还需完善下）
git-svn-id: http://code.taobao.org/svn/tfs@2431 b722c2b5-6a69-4643-bcd1-1a75825fa998

+++ branches/dev_from_stable2.5_for_lifecycle/src/common/expire_define.h
@@ -68,7 +68,7 @@ namespace tfs
                                 char *data, const int32_t size);
       static int deserialize_es_stat_key(const char *data, const int32_t size,
                                   uint64_t *es_id, int32_t* es_num,
-                                  int32_t* task_time, int32_t *hash_bucket_num, int32_t *sum_file_num);
+                                  int32_t* task_time, int32_t *hash_bucket_num, int64_t *sum_file_num);
 
       static int transfer_time(const int32_t time, int32_t *p_days_secs, int32_t *p_hours_secs);
     };
